# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Purchase Orders API",
        "summary": "The upload legacy application and queries.",
        "description": "This is an api that consumes legacy files to database and provides endpoits to consulting them.",
        "termsOfService": "https://example.com/terms/",
        "contact": {
          "name": "API Support",
          "url": "https://github.com/MatheusMaricondi/legacy-file-handler",
          "email": "matheus.maricondi@gmail.com"
        },
        "license": {
          "name": "Apache 2.0",
          "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.1"
      },
      "servers": [
        {
          "url": "http://localhost:3000/v1",
          "description": "API of queries"
        }
      ],
      "paths": {
        "/orders": {
          "post": {
            "summary":"Find orders by range date",
            "description": "This endpoint bring the order a range of date",
            "tags": ["Order by date"],
            "requestBody": {
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schema/DateRange"
                  },
                  "examples": {
                    "orders asc": {
                      "value": {
                        "initDate": "2021-05-10",
                        "endDate": "2023-11-15",
                        "order": "asc"
                      }
                    },
                    "orders desc": {
                      "value": {
                        "initDate": "2021-05-10",
                        "endDate": "2023-11-15",
                        "order": "desc"
                      }
                    }
                  }
                }
              }
            },
            "responses": {
              "500": {
                "description": "Server error"
              },
              "200": {
                "description": "OK",
                "content": {
                  "application/json": {
                    "schema": {
                      "items": {
                        "$ref": "#/components/schema/DateRange"
                      }
                    }
                  }
                
                }
              }
            }
          }
        },
        "/orders/{id}": {
          "get": {
            "summary":"Find orders by id",
            "description": "This endpoint bring the order by id",
            "tags": ["Order by id"],
            "parameters": [{
              "name": "id",
              "in": "path",
              "description": "Order name by id",
              "required": true
            }],
            "responses": {
              "500": {
                "description": "Server error"
              },
              "200": {
                "description": "OK",
                "content": {
                  "application/json": {
                    "schema": {
                      "items": {
                        "$ref": "#/components/schema/DateRange"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "/upload": {
          "get": {
            "summary":"Upload files",
            "description": "This endpoint receive a text file and save them in database",
            "tags": ["Uploads"],
            "requestBody": {
              "content": {
                "multipart/form-data": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "filename": {
                        "type": "string",
                        "format": "binary"
                      }
                    }
                  }
                }
              }
            },
            "responses": {
              "500": {
                "description": {"error": "Server error"}
              },
              "400": {
                "description": {"message": "Invalid file type"}
              },
              "201": {
                "description": {"message": "clients file was sucessfully uploaded", "file": "#/components/schema/Upload"},
                "content": {
                  "application/json": {
                    "schema": {
                      "items": {
                        "$ref": "#/components/schema/Upload"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "components": {
        "schema": {
          "DateRange": {
            "type": "object",
            "properties":{
              "initDate": {"type": "string"},
              "endDate": {"type": "string"},
              "order": {"type": "string"}
            }
          },
          "Upload": {
            "type": "object",
            "properties":{
              "message": {"type": "string"},
              "file": {
                "type": "object",
                "properties":{
                  "fieldname": {"type": "string"},
                  "originalname": {"type": "string"},
                  "encoding": {"type": "string"},
                  "mimetype": {"type": "string"},
                  "destination": {"type": "string"},
                  "filename": {"type": "string"},
                  "path": {"type": "string"},
                  "size": {"type": "number"}
                }
              }
            }
          }
        }
      }
}